Moved paired/unpaired getter to its own class that handles RDD caching properly, fixed mask parser test error
