+++ src/test/java/org/broadinstitute/hellbender/tools/spark/pathseq/PSUtilsTest.java
@@ -2,9 +2,6 @@ package org.broadinstitute.hellbender.tools.spark.pathseq;
 
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.io.Input;
-import htsjdk.samtools.SAMFileHeader;
-import htsjdk.samtools.SAMSequenceDictionary;
-import htsjdk.samtools.SAMSequenceRecord;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.broadinstitute.hellbender.engine.spark.SparkContextFactory;
@@ -16,12 +13,10 @@ import org.broadinstitute.hellbender.utils.test.BaseTest;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-import scala.Tuple2;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 
 public class PSUtilsTest extends BaseTest {
@@ -35,7 +30,6 @@ public class PSUtilsTest extends BaseTest {
     public Object[][] getAlignmentData() {
         return new Object[][]{
                 {"", 0, new byte[0], Boolean.FALSE},
-                {"0", 31, null, Boolean.TRUE},
                 {"0,1", 31, new byte[]{0, 1}, Boolean.FALSE},
                 {"0,1,", 31, new byte[]{0, 1}, Boolean.FALSE},
                 {"0,1,10,8", 31, new byte[]{0, 1, 10, 8}, Boolean.FALSE},
@@ -81,68 +75,6 @@ public class PSUtilsTest extends BaseTest {
         Assert.assertTrue(result.contains(primaryRead));
         Assert.assertTrue(result.size() == 1);
 
-    }
-
-    @Test(groups = "spark")
-    public void testReadPairing() {
-        final JavaSparkContext ctx = SparkContextFactory.getTestSparkContext();
-        final List<GATKRead> readList = new ArrayList<>(2);
-
-        final SAMSequenceDictionary seq = new SAMSequenceDictionary();
-        seq.addSequence(new SAMSequenceRecord("test_seq", 1000));
-        final SAMFileHeader header = new SAMFileHeader(seq);
-
-        final int numReadPairs = 300;
-        final int numUnpairedReads = 205;
-        final int numFormerlyPairedReads = 400;
-        final List<String> pairedReadNames = new ArrayList<>(numReadPairs * 2);
-        final List<String> unpairedReadNames = new ArrayList<>(numUnpairedReads + numFormerlyPairedReads);
-
-        for (int i = 0; i < numReadPairs; i++) {
-            final String readName = "paired_" + i;
-            final List<GATKRead> readPair = ArtificialReadUtils.createPair(header, readName, 101, 1, 102, false, false);
-            readList.addAll(readPair);
-            pairedReadNames.add(readName);
-            pairedReadNames.add(readName);
-        }
-
-        for (int i = 0; i < numUnpairedReads; i++) {
-            final String readName = "unpaired_" + i;
-            final GATKRead unpairedRead = ArtificialReadUtils.createRandomRead(101);
-            unpairedRead.setName(readName);
-            readList.add(unpairedRead);
-            unpairedReadNames.add(readName);
-        }
-
-        for (int i = 0; i < numFormerlyPairedReads; i++) {
-            final String readName = "formerly_paired_" + i;
-            final List<GATKRead> readPair = ArtificialReadUtils.createPair(header, readName, 101, 1, 102, false, false);
-            final GATKRead formerlyPairedRead = readPair.get(0);
-            readList.add(formerlyPairedRead);
-            unpairedReadNames.add(readName);
-        }
-
-        Collections.shuffle(readList);
-        final int numPartitions = 3;
-        final JavaRDD<GATKRead> readRDD = ctx.parallelize(readList,numPartitions);
-        final int readsPerPartitionGuess = (numReadPairs + numUnpairedReads + numFormerlyPairedReads) / numPartitions;
-        final Tuple2<JavaRDD<GATKRead>,JavaRDD<GATKRead>> result = PSUtils.splitPairedAndUnpairedReads(readRDD, readsPerPartitionGuess);
-
-        final List<GATKRead> resultPaired = result._1.collect();
-        Assert.assertEquals(resultPaired.size(), pairedReadNames.size());
-        for (final GATKRead read : resultPaired) {
-            final String readName = read.getName();
-            Assert.assertTrue(pairedReadNames.contains(readName));
-            pairedReadNames.remove(readName);
-        }
-
-        final List<GATKRead> resultUnpaired = result._2.collect();
-        Assert.assertEquals(resultUnpaired.size(), unpairedReadNames.size());
-        for (final GATKRead read : resultUnpaired) {
-            final String readName = read.getName();
-            Assert.assertTrue(unpairedReadNames.contains(readName));
-            unpairedReadNames.remove(readName);
-        }
     }
 
     @Test
